requirements.txt
# OpenAI CLIP (matches freeze.txt commit)
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1

# PyTorch ecosystem
torch>=2.5.0
torchvision>=0.20.0

# Image and CLI utilities
pillow>=11.3.0
tabulate>=0.9.0
icrawler>=0.6.10
tqdm
